backend/realestate_agent/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
transformers<4.50
nltk>=3.8.1
pyttsx3>=2.90
python-multipart>=0.0.5
python-dotenv>=0.19.0
pydantic>=1.8.0
